put screen back
I reconsidered and I am putting screen back to see if I can use to better track analyses at the command line in the container. The reasoning is that even though attaching and detaching container works for keeping running when started with `-itd` flags, I lose the record of everything that happened while detached, as well as prior to that.