add streamlined version & make more self-conatined
add streamlined version to make docker image smaller and dealing with genome files so that the large download to get the two relatively small files needed for mapping the reads is handled and done in a way that the large download doesn't add to size of the Docker image. Plus the addition of the two files means it is more self-contained and will work even if those files go away.